initial upload
Used to merge *.geotif files with exactly the same structure to a single one.  It was used to combine the geoteiffs of the individual profiles from a survey into a complete overview raster. The No data value, which should later be set transparently in QGIS, can be set individually (in the script, default set to -99999)